utils/requirements.txt
paho-mqtt==2.1.0
pillow==11.2.1
PyYAML==6.0.2
